make the task's description more clear
Changed description in line 12 from:
'A string can start with a letter or, for example, a dot or space.'
to:
'A string can start with a letter or, for example, one/multiple dot(s) or space(s)'
in order to Correspond to the string input: ' ... and so on ... ' during the checking process, which has multiple dots leading the text, rather than what ' a dot or space ' described.
P.S. English is not my mother language so it's quite confusing for me to predict the input content with the previous instruction. Same time, I'll be sorry if there's a mistake in my description.